+++ test/core/detail/variant.cpp
@@ -9,6 +9,7 @@
 #include <boost/beast/core/detail/variant.hpp>
 
 #include <boost/beast/unit_test/suite.hpp>
+#include <exception>
 #include <string>
 
 namespace boost {
@@ -29,17 +30,44 @@ public:
             return n;
         }
 
+        int value = 0;
         bool move = false;
         bool copy = false;
-        bool move_assign = false;
-        bool copy_assign = false;
 
-        Q() { ++count(); }
-        ~Q() { --count(); }
-        Q(Q&&) { ++count(); move = true; }
-        Q(Q const&) { ++count(); copy = true; }
-        Q& operator=(Q&&) { move_assign = true; return *this; }
-        Q& operator=(Q const&) { copy_assign = true; return *this; }
+        Q& operator=(Q&& q) = delete;
+        Q& operator=(Q const& q) = delete;
+
+        ~Q()
+        { 
+            --count();
+        }
+
+        Q()
+        {
+            ++count();
+        }
+                
+        Q(Q&& q)
+        {
+            if(q.value < 0)
+                throw std::exception{};
+            ++count();
+            move = true;
+        }
+
+        Q(Q const& q)
+        {
+            if(q.value < 0)
+                throw std::exception{};
+            ++count();
+            copy = true;
+        }
+
+        explicit Q(int value_)
+            : value(value_)
+        {
+            ++count();
+        }
     };
 
     void
@@ -58,6 +86,9 @@ public:
             v.emplace<1>(10);
             BEAST_EXPECT(v.index() == 1);
             BEAST_EXPECT(v.get<1>() == 10);
+            v.emplace<1>(20);
+            BEAST_EXPECT(v.index() == 1);
+            BEAST_EXPECT(v.get<1>() == 20);
         }
         {
             variant<std::string> v;
@@ -70,6 +101,10 @@ public:
             BEAST_EXPECT(Q<1>::count() == 0);
             v.emplace<1>();
             BEAST_EXPECT(Q<1>::count() == 1);
+            BEAST_EXPECT(v.get<1>().value == 0);
+            v.emplace<1>(42);
+            BEAST_EXPECT(Q<1>::count() == 1);
+            BEAST_EXPECT(v.get<1>().value == 42);
             v.reset();
             BEAST_EXPECT(Q<1>::count() == 0);
         }
@@ -201,39 +236,50 @@ public:
             BEAST_EXPECT(v3.get<1>() == "Hello");
         }
         {
+            variant<Q<1>, Q<2>> v1;
+            variant<Q<1>, Q<2>> v2;
+            v1.emplace<1>();
+            BEAST_EXPECT(v1.index() == 1);
+            BEAST_EXPECT(Q<1>::count() == 1);
+            v2.emplace<2>();
+            BEAST_EXPECT(v2.index() == 2);
+            BEAST_EXPECT(Q<2>::count() == 1);
+            v2 = std::move(v1);
+            BEAST_EXPECT(v1.index() == 0);
+            BEAST_EXPECT(v2.index() == 1);
+            BEAST_EXPECT(Q<1>::count() == 1);
+            BEAST_EXPECT(Q<2>::count() == 0);
+        }
+        BEAST_EXPECT(Q<1>::count() == 0);
+        BEAST_EXPECT(Q<2>::count() == 0);
+        {
             variant<Q<1>> v1;
             v1.emplace<1>();
-            BEAST_EXPECT(! v1.get<1>().move_assign);
             BEAST_EXPECT(Q<1>::count() == 1);
             variant<Q<1>> v2;
             v2.emplace<1>();
             BEAST_EXPECT(Q<1>::count() == 2);
-            BEAST_EXPECT(! v2.get<1>().move_assign);
             v2 = std::move(v1);
             BEAST_EXPECT(v1.index() == 0);
             BEAST_EXPECT(v2.index() == 1);
             BEAST_EXPECT(v2.get<1>().move);
-            BEAST_EXPECT(!v2.get<1>().move_assign);
             BEAST_EXPECT(Q<1>::count() == 1);
         }
         BEAST_EXPECT(Q<1>::count() == 0);
         {
             variant<Q<1>, Q<2>, Q<3>> v1;
             v1.emplace<2>();
-            BEAST_EXPECT(! v1.get<2>().move_assign);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 1);
             BEAST_EXPECT(Q<3>::count() == 0);
             variant<Q<1>, Q<2>, Q<3>> v2;
             v2.emplace<2>();
-            BEAST_EXPECT(! v2.get<2>().move_assign);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 2);
             BEAST_EXPECT(Q<3>::count() == 0);
             v2 = std::move(v1);
             BEAST_EXPECT(v1.index() == 0);
             BEAST_EXPECT(v2.index() == 2);
-            BEAST_EXPECT(! v2.get<2>().move_assign);
             BEAST_EXPECT(v2.get<2>().move);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 1);
@@ -258,18 +304,30 @@ public:
             BEAST_EXPECT(v3.get<1>() == "Hello");
         }
         {
+            variant<Q<1>, Q<2>> v1;
+            variant<Q<1>, Q<2>> v2;
+            v1.emplace<1>();
+            BEAST_EXPECT(v1.index() == 1);
+            BEAST_EXPECT(Q<1>::count() == 1);
+            v2.emplace<2>();
+            BEAST_EXPECT(v2.index() == 2);
+            BEAST_EXPECT(Q<2>::count() == 1);
+            v2 = v1;
+            BEAST_EXPECT(v1.index() == 1);
+            BEAST_EXPECT(v2.index() == 1);
+            BEAST_EXPECT(Q<1>::count() == 2);
+            BEAST_EXPECT(Q<2>::count() == 0);
+        }
+        {
             variant<Q<1>> v1;
             v1.emplace<1>();
-            BEAST_EXPECT(! v1.get<1>().copy_assign);
             BEAST_EXPECT(Q<1>::count() == 1);
             variant<Q<1>> v2;
             v2.emplace<1>();
             BEAST_EXPECT(Q<1>::count() == 2);
-            BEAST_EXPECT(! v2.get<1>().copy_assign);
             v2 = v1;
             BEAST_EXPECT(v1.index() == 1);
             BEAST_EXPECT(v2.index() == 1);
-            BEAST_EXPECT(! v2.get<1>().copy_assign);
             BEAST_EXPECT(v2.get<1>().copy);
             BEAST_EXPECT(Q<1>::count() == 2);
         }
@@ -277,20 +335,17 @@ public:
         {
             variant<Q<1>, Q<2>, Q<3>> v1;
             v1.emplace<2>();
-            BEAST_EXPECT(! v1.get<2>().copy_assign);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 1);
             BEAST_EXPECT(Q<3>::count() == 0);
             variant<Q<1>, Q<2>, Q<3>> v2;
             v2.emplace<2>();
-            BEAST_EXPECT(! v2.get<2>().copy_assign);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 2);
             BEAST_EXPECT(Q<3>::count() == 0);
             v2 = v1;
             BEAST_EXPECT(v1.index() == 2);
             BEAST_EXPECT(v2.index() == 2);
-            BEAST_EXPECT(! v2.get<2>().copy_assign);
             BEAST_EXPECT(v2.get<2>().copy);
             BEAST_EXPECT(Q<1>::count() == 0);
             BEAST_EXPECT(Q<2>::count() == 2);
@@ -340,6 +395,48 @@ public:
         BEAST_EXPECT(Q<1>::count() == 0);
         BEAST_EXPECT(Q<2>::count() == 0);
         BEAST_EXPECT(Q<3>::count() == 0);
+        
+        // basic guarantee
+        {
+            // move
+            variant<Q<1>> v1;
+            v1.emplace<1>();
+            BEAST_EXPECT(Q<1>::count() == 1);
+            variant<Q<1>> v2;
+            v2.emplace<1>(-1);
+            BEAST_EXPECT(Q<1>::count() == 2);
+            try
+            {
+                v1 = std::move(v2);
+                fail();
+            }
+            catch(std::exception const&)
+            {
+                BEAST_EXPECT(v1.index() == 0);
+                BEAST_EXPECT(Q<1>::count() == 1);
+            }
+        }
+        BEAST_EXPECT(Q<1>::count() == 0);
+        {
+            // copy
+            variant<Q<1>> v1;
+            v1.emplace<1>();
+            BEAST_EXPECT(Q<1>::count() == 1);
+            variant<Q<1>> v2;
+            v2.emplace<1>(-1);
+            BEAST_EXPECT(Q<1>::count() == 2);
+            try
+            {
+                v1 = v2;
+                fail();
+            }
+            catch(std::exception const&)
+            {
+                BEAST_EXPECT(v1.index() == 0);
+                BEAST_EXPECT(Q<1>::count() == 1);
+            }
+        }
+        BEAST_EXPECT(Q<1>::count() == 0);
     }
 
     void