+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 87:
 
 * Update appveyor for Boost branch
 * Rename to BEAST_EXPECT
+* variant fixes and tests
 
 --------------------------------------------------------------------------------